+++ 2022/01/pan93412/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "pan93412"
+name = "pan93412_aoc_2022"
 version = "0.1.0"
 edition = "2021"
