+++ 2022/01/pan93412/tests/testdata.txt
@@ -1,0 +1,14 @@
+1000
+2000
+3000
+
+4000
+
+5000
+6000
+
+7000
+8000
+9000
+
+10000
